move font methods to ScottPlot.Config.Fonts module
